s3_store: modify KVS layout for S3, add 'name' key
Every S3 operation on objects require the object full path.
The operation complexity, unless I missed something, depends at least on
the number of entries in every directory from the object up to the root.
This is probably the case as the full path is rarely (never?) required
for other stores.
By adding a "[inode].name" key for each entry, containing the name,
getting the full path now depends on the number of directories from the
object to the root.

NOTE:
The code that addes/removes this new key lies in libkvsns, and so is
guarded by a new KVSNS_S3 #define, manually declared in kvsns.h for now.

If this solution is to be kept, a cleaner way of defining KVSNS_S3
should be thought of.