s3_store: implement delete_object [WIP]
[WIP]

In current implementation, in which s3 extstore adds a new key to redis
(i.e the X.name key), extstore_del can't build the S3 object path
because the metadata has already been removed from redis when
extstore_del gets called.

One way (outside of the KVS) to retrieve the object S3 full path from
its ino is required. A quick and dirty solution would be to store the
whole path in a map ino_t => fullpath.

A more scalable solution would be to have the full path stored in a
prefix trie, and a way to retrieve the trie node from the object inode.
Reconstructing the path from the trie node would be easy.