Set newly uploaded file caller's uid/gid
Uid/Gid were previously wrong for newly uploaded file, because at the
moment we perform the upload (`extstore_close`), we don't know any more
who created the file, caller's uid and gid were accessible in
`extstore_create` but are lost later.

The `trick` here it to set uid and gid of the "write cache file" when
we create it in ``extstore_create` so that we just have to retrieve both
values and use them in the `stat` struct used for upload to s3
(`put_object`).